fix(gui): reserve recorder-cluster and icon slots in Shortcuts rows (#264)
Unbound Shortcuts rows showed a left gap before the record
button. Two reserved-width slots that weren't actually reserved:

1. `KeyRecorderField`'s trailing cluster (conflict badge, record
   button, clear "×") only reserved the record button's width, so
   the cluster's total width varied by state and the row's
   `Spacer` pushed the narrowest (unbound) state furthest right —
   maximizing the label↔button gap exactly when unbound. A shared
   local `iconSlot` helper reserves a fixed slot for the badge and
   the clear button (a `Color.clear` base with the glyph in an
   overlay, so an absent glyph still holds the width), pinning the
   record button's leading edge across unbound/bound/conflict.

2. `NavRow`'s icon block rendered conditionally — only Go-to and
   Move-to-space rows carry a glyph — so every icon-less row's
   label started ~18 pt further left. Reserve the 18 pt slot
   unconditionally the same way.

`NavRow` is extracted to its own file (`KeybindingNavRow.swift`):
the added slot pushed `KeybindingSections` to 346/350, so the
shared row every section composes moves out to keep both files
well under the ceiling (§5 split-early).

Pure layout; no behavior, data, or copy change.
