+++ Sources/KiwiDesk/Settings/Tabs/KeybindingSections.swift
@@ -218,119 +218,3 @@ struct ChangeModesSection: View {
         modeNames.filter { $0 != current }
     }
 }
-
-/// A single catalog row: label plus a recorder that upserts a
-/// `.navigation` binding keyed by the command's Lua. The
-/// preflight hard-blocks combos already held by another row
-/// (#34); the row id feeds "Go to" (#68 §3.6.2).
-struct NavRow: View {
-    @ObservedObject var model: SettingsModel
-    @Binding var bindings: [KeyBinding]
-    let command: NavCommand
-    @Environment(\.keybindingOverrideBase)
-    private var overrideBase
-    @Environment(\.keybindingModeName)
-    private var modeName
-
-    var body: some View {
-        HStack {
-            if let glyph = command.icon, !glyph.isEmpty {
-                IconGlyphLabel(icon: glyph)
-                    .frame(width: 18)
-            }
-            Text(command.resolvedLabel)
-            Spacer()
-            KeyRecorderField(
-                combo: index.map { bindings[$0].combo } ?? "",
-                conflict: conflict,
-                preflight: preflight,
-                onRecord: record,
-                onClear: clear
-            )
-        }
-        .keybindingRowStyle(inherited: isInherited)
-        .id(command.lua)
-    }
-
-    private var index: Int? {
-        bindings.firstIndex {
-            $0.kind == .navigation && $0.lua == command.lua
-        }
-    }
-
-    /// Override mode: bound-and-equal to the base row, or
-    /// unbound on both sides. Always false while editing live.
-    private var isInherited: Bool {
-        guard let base = overrideBase else { return false }
-        if let index {
-            return bindings[index].isInherited(from: base)
-        }
-        return !base.contains {
-            $0.kind == .navigation && $0.lua == command.lua
-        }
-    }
-
-    private var conflict: String? {
-        guard let index else { return nil }
-        return KeybindingConflicts.text(
-            for: bindings[index],
-            in: bindings
-        )
-    }
-
-    private func preflight(
-        _ combo: String
-    ) -> RecorderRejection? {
-        RecorderPreflight.rejection(
-            combo: combo,
-            excluding: { [lua = command.lua] in
-                $0.kind == .navigation && $0.lua == lua
-            },
-            bindings: $bindings,
-            // Steal live-applies too (via `record`); only the
-            // recorder's own commit shows the caption.
-            commit: { _ = record($0) }
-        )
-    }
-
-    private func record(
-        _ combo: String
-    ) -> LiveApplyFeedback? {
-        if let index {
-            bindings[index].combo = combo
-        } else {
-            bindings.append(
-                KeyBinding(
-                    combo: combo,
-                    lua: command.lua,
-                    kind: .navigation,
-                    label: command.label
-                )
-            )
-        }
-        if let updated = bindings.first(where: {
-            $0.kind == .navigation && $0.lua == command.lua
-        }) {
-            model.noteRecordedCombo(updated, in: bindings)
-            return model.liveApplyRecorded(
-                modeName: modeName,
-                bindingID: updated.id,
-                combo: combo
-            )
-        }
-        return nil
-    }
-
-    private func clear() {
-        guard let index else { return }
-        let id = bindings[index].id
-        bindings.remove(at: index)
-        // Live target: the removed hotkey unregisters now
-        // (#123); no caption for a clear.
-        _ = model.liveApplyRecorded(
-            modeName: modeName,
-            bindingID: id,
-            combo: nil
-        )
-    }
-}
